requirements.txt
pyyaml==5.4
numpy==1.19.5
category_encoders==2.1.0
kmodes==0.10.2
xgboost==1.3.3
tqdm==4.40.2
imbalanced_learn==0.6.1
pandas==0.25.3
dill==0.3.8
lime==0.1.1.37
scipy==1.4.1
tensorflow_gpu==2.4.0
tensorboard==2.4.1
matplotlib==3.3.3
joblib==0.14.1
scikit-learn==0.22.1
h5py==2.10.0
protobuf==3.20
pytest==7.0.1